requirements.txt
tqdm
PyMuPDF
pandas
uvicorn
fastapi
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
google-cloud-storage>=2.13.0
openai
pinecone
dotenv
python-multipart
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-grpc